axiomflow-api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.4.0
SQLAlchemy>=2.0.30
PyMySQL>=1.1.1
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0
email-validator>=2.2.0
aiosmtplib>=3.0.1
jinja2>=3.1.4
python-multipart>=0.0.9
slowapi>=0.1.9
Pillow>=10.4.0
